修复不支持逻辑表达式和算术表达式混合的bug

+++ src/main/kotlin/cn/yumetsuki/semantic_analysis/ASTVisitor.kt
@@ -86,7 +86,7 @@ class DefaultASTVisitor(
     }
 
     override fun visitNumber(node: NumberNode): Number {
-        return node.value
+        return node.value.toDouble()
     }
 
     override fun visitVariableDefineAndAssignment(node: VariableDefineAndAssignmentNode): Pair<String, Any> {
